bot_v1/requirements.txt
MetaTrader5
pandas
numpy
pyyaml
requests
python-dotenv
